wip: refactor types
