+++ playgrounds/vite-svelte/package.json
@@ -17,8 +17,6 @@
     "@tsconfig/svelte": "^4.0.1",
     "svelte": "^3.58.0",
     "svelte-check": "^3.3.1",
-    "tslib": "^2.5.0",
-    "typescript": "^5.0.2",
-    "vite": "^4.3.9"
+    "tslib": "^2.5.0"
   }
 }
