+++ playgrounds/vite-vue/package.json
@@ -14,8 +14,6 @@
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^4.1.0",
-    "typescript": "^5.0.2",
-    "vite": "^4.3.9",
     "vue-tsc": "^1.4.2"
   }
 }
